+++ lib/models/expenditure.dart
@@ -8,7 +8,7 @@ class Expenditure {
   String name = '';
   ExpenditureFrequency frequency = ExpenditureFrequency.Monthly;
   int amount = 0;
-  int yearlyAppreciationPercentage = 0;
+  double yearlyAppreciationPercentage = 0;
   Jiffy startDate = Jiffy().startOf(Units.DAY);
   Jiffy endDate = Jiffy().startOf(Units.DAY).add(years: 100);
   bool belongsToAsset = false;
